requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-allauth==0.63.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-ninja==1.2.2
djangorestframework==3.15.2
Faker==26.0.0
h11==0.14.0
idna==3.7
oauthlib==3.2.2
outcome==1.3.0.post0
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
